Fixed Wall Distance
I wasn't sure where you were going with the dot product, so I ended up
using a way that made more sense to me. Feel free to propogate my
changes to the much cleaner dot-product calculation if you can get it to
work.

My method blindly calculates the pythogorean distance to the intercept
from the robot, then resets to 9999 if we find out that the interception
point isn't actually on the wall or it intersects behind the laser
scanner rather than in front of it.

All of these tests are done using various pythogorean distance
calculations. Not elegant, but since the robot isn't very
code-performance reliant, it should be alright?

David